add helper functions for getting mp id/area info fixed phantom unshackle to work with remaster also re-enable unrestrict_network_synced_effectids (forgot to when i fixed it)

+++ OverhaulDLL/src/AnimationEdits.cpp
@@ -71,7 +71,7 @@ static const std::unordered_map<int32_t, std::tuple<float, float>> ANIMATIONS_TO
 
 
 extern "C" {
-    uint32_t animation_entry_set_return;
+    uint64_t animation_entry_set_return;
     void animation_entry_set_injection();
     void read_body_aid_injection_helper_function(int32_t*, float*);
 }
